schemas: Consistently use child schemas
Currently subschemas in org.gnome.mutter are marked as child schemas,
but subschemas in org.gnome.mutter.wayland are not. There's no harm
with that, it's just nice to be consistent, and having schemas show
up in `gsettings list-children` is a nice touch.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2809>