readme: Expand how to tag commit messages
It's a common mistake that links to issues being resolved are not URLs,
and that Part-of was added manually by contributors. Make it clear what
the expectations are regarding this. Formalize on `Fixes:` for fixing
commit messages, and `Closes:` for closing issues while at it. Format
and alias for generating `Fixes:` tags come from mesa.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2847>